Update mydoc_3Dculture.md
Daniel, at the step of harvesting the cells, I used to adjust the volume to 15 ml and used as needed for microscopic or RNA Seq experiments. It worked well but,  I believe this may be adjusted, depending on the concentration of cells targeted, if it is to be used again in experiments with bacteria. I had a hard time to count the cells, you remember. I would either have "clumps" or I had bare beads and damaged cells that were hard count anyway.